pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "retrieval-os"
version = "0.1.0"
description = "Multimodal Retrieval-Native Inference Runtime"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
    # Config & validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Cache
    "redis[hiredis]>=5.2.0",
    # Object storage
    "boto3>=1.35.0",
    # Vector index
    "qdrant-client>=1.12.0",
    # Observability
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "prometheus-client>=0.21.0",
    # Logging
    "structlog>=24.4.0",
    # HTTP client
    "httpx>=0.27.0",
    # Retry
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "anyio[trio]>=4.6.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]
ml = [
    "sentence-transformers>=3.0.0",
    "open-clip-torch>=2.24.0",
    "faster-whisper>=1.0.0",
    "timm>=1.0.0",
    "Pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/retrieval_os"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
